backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic[email]>=2.6
numpy
opencv-python
ultralytics
groq
python-dotenv
